requirements.txt
google-auth==2.22.0
pandas==2.0.3
pandas-gbq==0.19.2
pybaseball==2.2.7